Fixed some issues with using int instead of double. looks okay now.

+++ src/structure/Cell.java
@@ -96,7 +96,7 @@ public class Cell {
 		value = form.formula(formulaWithoutCellReference);
 	}
 	
-	public void setValue(int v){
+	public void setValue(double v){
 		value = String.valueOf(v);
 	}
